Respect TORCH_DISABLE_ADDR2LINE in symbolizer
If TORCH_DISABLE_ADDR2LINE is set, the symbolizer will instead give the
filename of the shared library as the filename, the offset in that library as the linenumber,
and use dladdr to get the function name if possible. This is much faster than using addr2line,
and the symbols can be later resolved offline using addr2line if desired.

ghstack-source-id: 6e4d2a4c887cea88e940a4ab8bf78bb4306ac499
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121359