[WIP] Add torch.cond support to AOT Inductor
Summary: In this PR, `torch.cond` support and the necessary
codegening infrastructure is added to C++ wrapper (AOTInductor and
friends).

N.B.: When run jointly, 2 of the added 42 new unit test configurations
are still failing with accuracy issues. When run separately, they don't.
I'm still debugging this, but the PR is otherwise ready for a review.

Test Plan:

```
$ python test/inductor/test_aot_inductor.py -k test_cond
...
----------------------------------------------------------------------
Ran 42 tests in 249.690s

FAILED (failures=2)
```

ghstack-source-id: 341781ef091fa2c25339b0d6c9b6ddb2d08e1780
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121120